config/requirements_backend.txt

# BestStag v9.1 Backend - Requirements Específicos
# Dependências adicionais para o backend FastAPI

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0
alembic>=1.12.0

# Async support
asyncpg>=0.29.0
aiofiles>=23.2.1

# Monitoring e Health Checks
psutil>=5.9.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation
email-validator>=2.1.0

# Background tasks
celery>=5.3.0
redis>=4.5.0

# HTTP client
httpx>=0.25.0

# JSON handling
orjson>=3.9.0

# Environment
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Testing
pytest-asyncio>=0.21.0
httpx>=0.25.0
